refactor: share prTooltip across all PR display surfaces
The per-check tooltip lived privately in RowIcons (dock pip). Tile
title bar, workspace switcher, and close-confirm dialog all hung a
bare \`prLabel\` tooltip on their PR icons — missing the check
verdict and per-check list that the dock already showed.

Extract \`prTooltip\` to \`packages/client/src/terminal/prTooltip.ts\`
(client-side; the \`kolu-github/schemas\` package stays type-only,
no UI helpers there). Four surfaces now read from the same
function — one verdict format everywhere a PR icon shows:

- Dock pip (cards + mobile drawer, via RowIcons)
- Tile title bar (TerminalMeta — both compact and full layouts)
- Workspace switcher card (WorkspaceGrid)
- Close-confirm dialog (CloseConfirm) — gained a tooltip it
  didn't have

The right-panel MetadataInspector is the one surface that doesn't
get a tooltip — it has the real estate to lay the per-check list
out inline as proper UI, so a fail/pending list is scannable
without hovering. Renders \`<For each={pr().checkRuns}>\` with
ChecksIndicator + check name per row under a new "Checks" Row.
